Bring contribution contributor logic in line with affiliation institution logic (similarly large dataset but runs faster)
